feat: Add Angel One broker, execution engine, WebSocket server, and wire frontend to backend
- Angel One SmartAPI broker adapter with TOTP auth
- Execution engine: market data → strategy evaluation → auto order placement
- Broker auth routes (login/logout/status/refresh)
- Engine control routes (start/stop/emergency-stop/reload)
- WebSocket server for real-time ticks, signals, and order updates
- Broker factory for managing active broker instance
- Frontend API URLs updated from port 8080 to 3001
- Fixed double /api/api/ bug in engine.ts and marketData.ts
- Replaced mock WebSocket with real WS connection (fallback to simulation)
- Fixed SQLite compatibility: BigInt→Int, JSON.stringify for String fields
- SQLite database with Prisma schema (no PostgreSQL needed)

+++ algo-trading/algo-trading-backend-node/src/services/marketDataService.ts
@@ -294,7 +294,7 @@ export class MarketDataService extends EventEmitter {
                 high: builder.high,
                 low: builder.low,
                 close: builder.close,
-                volume: BigInt(Math.floor(builder.volume)),
+                volume: Math.floor(builder.volume),
                 timestamp: builder.startTime,
             };
 
@@ -309,7 +309,7 @@ export class MarketDataService extends EventEmitter {
                         high: builder.high,
                         low: builder.low,
                         close: builder.close,
-                        volume: BigInt(Math.floor(builder.volume)),
+                        volume: Math.floor(builder.volume),
                     },
                 });
             } catch (error: any) {
